BETA 3.0

+++ CloudWeb/Models/db.cs
@@ -1,20 +1,22 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using CloudWeb.Models;
+using static CloudWeb.Models.Product;
 
 namespace CloudWeb.Models
 {
     public class db : DbContext
     {
         public static bool value = true;
-        public db(DbContextOptions<db> options) : base(options) {
+        public db(DbContextOptions<db> options) : base(options)
+        {
 
-            /*if (value)
+            if (value)
             {
                 Database.EnsureDeleted();
                 value = false;
             } 
-            Database.EnsureCreated();*/
-            
+            Database.EnsureCreated();
+
         }
         public DbSet<Company> Company { get; set; }
         public DbSet<Employee> Employee { get; set; }
@@ -22,5 +24,20 @@ namespace CloudWeb.Models
         public DbSet<Product> Product { get; set; }
         public DbSet<Order> Order { get; set; }
         public DbSet<Team> Team { get; set; }
+
+       protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<Company>().HasData(
+                new Company
+                {
+                    id = 1,
+                    nameCompany = "WevStorm",
+                    Location = "Tallinn",
+                    directorName = "Arnold",
+                    directorSurname = "Aghh",
+                    productType = (Company.branch)productType.Website
+                }
+            );
+        }
     }
 }
